إنقاذ المصابين بوباء كوفيد-19
The current health crisis is placing a great burden on hospital systems, even the most advanced ones, with difficulties related especially to the provision of artificial respirators.

However, experts fear that this interest in the most advanced part of assisted breathing technology will overshadow a higher priority in the most vulnerable health systems: medical oxygen, which is a major component of intensive care units.

"The reality is that oxygen is the only life-saving treatment in Africa and the Asia-Pacific region at this point," says Hamish Graham, a pediatrician and researcher at the University of Melbourne.

"I am afraid that the excessive focus on respirators will be fatal if the oxygen problems are not resolved," he told AFP.

A report published in February of thousands of cases in China showed that nearly 20% of COVID-19 patients required oxygen.

The new Corona virus attacks the lungs, causing severe forms of respiratory distress, and also leads to a dangerous decrease in the level of oxygen in the blood.

Graham notes that "the presence of oxygen is self-evident in the hospitals of rich countries." But in the rest of the world, "therapists are very aware of the problem as they struggle every day to provide their patients with oxygen."

shortages in Africa and Asia
Many major hospitals in developing countries have oxygen bottles in operating rooms and medical services, as well as oxygen generators, which are portable equipment that allows purification of the surrounding air.

However, several studies show that less than half of hospitals in Africa and Asia-Pacific have oxygen available at any given time, Graham says. A smaller number of these institutions are equipped with pulse oximetry, which are small devices placed at the tip of a patient's finger to measure the level of oxygen in the blood.

This concern is not new for people treating patients with pneumonia, the most deadly infectious disease in children under five.

Despite the adopted domestic policies, the situation in some parts of Nigeria, one of the countries most affected by this situation, is "a very serious concern", according to Adamu Isah of the non-governmental organization Save the Children.

It is common to see children "suffering and suffocating", this former doctor told AFP. "We feel useless. We can't do much without oxygen."

'terrifying' situation

For her part, coordinator of the "Every Breath Counts" coalition, Leith Greenslade, explains that "health systems in Africa and South Asia were not prepared at all for such an epidemic and did not invest in respiratory treatments", describing the situation as "terrifying."

Although 800,000 children die of pneumonia annually, the disease does not receive the same attention as AIDS, malaria and tuberculosis.

The global health authorities have also "completely neglected" the issue of oxygen, according to Leith Greenslade, who notes that "the absence of global data (regarding oxygen supply) will constitute a major problem in responding to Covid-19 because we are going nowhere and we do not know which countries are most in need."

And she pointed out that "despite the fact that the epidemic is still relatively weak in Africa and some parts of Asia, there is likely to be a time margin of two months" for action.

Although it is difficult to predict how the virus will spread in this part of the world, experts are calling for emergency care systems to be strengthened.

“As everywhere, the (epidemic) curve should be flattened, but if institutions lack ICU beds, or have few, as in Malawi, which has 25 beds per 17 million people, says Gwen Haynes of Save the Children, It will not be possible to achieve the desired goal."